requirements.txt
pandas==2.3.0
geopandas==1.1.0
numpy==1.26.4
scipy==1.15.2
mlflow==2.21.0
statsmodels==0.14.4
fastapi==0.115.4
uvicorn==0.34.0
pydantic==2.10.6
scikit-learn==1.7.0
torch==2.6.0
torchvision==0.21.0
torch-geometric==2.6.1
torch_geometric_temporal==0.56.1
torch_sparse==0.6.18
torch_scatter==2.1.2
duckdb==1.3.1
shap==0.48.0
shapely==2.0.6
sktime==0.38.3
prophet==1.1.7
matplotlib==3.10.1
seaborn==0.13.2
geoplot==0.5.1
meteostat==1.7.4
xgboost==3.0.0
libpysal==4.12.1
esda==2.7.0
python-dotenv
boto3==1.35.53
botocore==1.35.99